Update rarefyBioTIME.R - CRAN corrections
\dontrun{} should only be used if the example really cannot be executed 
(e.g. because of missing additional software, missing API keys, ...) by 
the user. That's why wrapping examples in \dontrun{} adds the comment 
("# Not run:") as a warning for the user. Please replace \dontrun with 
\donttest where possible.

Please unwrap the examples if they are executable in < 5 sec, or replace 
dontrun{} with \donttest{}.

Left the example for the internal rarefysample for future work on the functions