fix tests to pass regression caused by changes in vegan
Vegan diversity() could drop names with tibbles. This was regarded
as a bug in vegan and fixed in 2.8-0. BioTIMEr test-metrics.R failed
because it tested that diversity results do not have site names.
There were no differences in diversity results, only in names.

This is a better fix than 0b6d4f7 to skip CRAN tests which still gives
error on github reverse dependency tests (and may give an error
in CRAN reverse dependency tests).