Made getAlphaMetrics faster
assertion making it clear that when measure = c(ABUNDANCE,BIOMASS) there is a helpful error message saying that only one can be used at a time. BUT the internals of the function actually handle both at the same time perfectly! This is blocked for consistency with getBetaMetrics and will be addressed in a future version